refactor(terminal): restructure autocompletion logic
